requirements.txt
requests
beautifulsoup4
mysql-connector-python
fastapi
uvicorn[standard]
